Refactor ATI params
