+++ src/app/components/ATIAnalytics/params/buildParams.test.ts
@@ -1,7 +1,7 @@
 import { TOPIC_PAGE } from '#app/routes/utils/pageTypes';
 import * as analyticsUtils from '#lib/analyticsUtils';
-import { RequestContextProps } from '../../../../contexts/RequestContext';
-import { ServiceConfig } from '../../../../models/types/serviceConfig';
+import { RequestContextProps } from '../../../contexts/RequestContext';
+import { ServiceConfig } from '../../../models/types/serviceConfig';
 import { buildPageATIParams, buildPageATIUrl } from './buildParams';
 
 (analyticsUtils.getAtUserId as jest.Mock) = jest.fn();
